draft of pipeline ready
